Simplify server startup configuration - eliminate hidden overrides
Consolidates server files to eliminate confusion about which server runs in production:

BEFORE:
- render.yaml claimed to run "node server.js"
- render-start.js secretly overrode and ran server-enterprise-complete.js
- Multiple conflicting server files caused deployment confusion
- Hidden overrides made it impossible to know what actually runs

AFTER:
- Single source of truth: server.js contains full enterprise functionality
- No hidden overrides or conflicting scripts
- What configs say is exactly what runs in production
- Clear, predictable server startup process

Changes:
- Rename server-enterprise-complete.js → server.js (now the main production server)
- Move legacy server files to archive/ folder for safety
- Remove render-start.js hidden override
- Update package.json to remove start:render script
- Update CLAUDE.md documentation with simplified configuration
- Preserve server/index.js for local development only

Configuration now transparent: render.yaml "node server.js" actually runs server.js
