+++ archive/server-basic.js
@@ -1,0 +1,990 @@
+import express from 'express';
+import path from 'path';
+import fs from 'fs';
+import { fileURLToPath } from 'url';
+
+// ES module equivalent of __dirname
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+// Import MCP client for live data integration
+let getMCPClient = null;
+try {
+  const mcpModule = await import('./services/mcp-client.js');
+  getMCPClient = mcpModule.getMCPClient;
+  console.log('✅ MCP Client imported successfully');
+} catch (error) {
+  console.warn('⚠️ MCP Client import failed:', error.message);
+}
+
+const app = express();
+const PORT = process.env.PORT || 10000;
+
+console.log('🚀 Starting Sentia Manufacturing Dashboard Server...');
+console.log('🌍 Environment:', process.env.NODE_ENV || 'development');
+
+// Middleware for parsing JSON
+app.use(express.json());
+
+// Try multiple paths to find the dist folder
+const possiblePaths = [
+  path.join(__dirname, 'dist'),
+  path.join(__dirname, '../dist'),
+  path.join(process.cwd(), 'dist'),
+  '/opt/render/project/src/dist'
+];
+
+let staticPath = null;
+for (const testPath of possiblePaths) {
+  console.log(`Checking for dist at: ${testPath}`);
+  try {
+    if (fs.existsSync(testPath)) {
+      staticPath = testPath;
+      console.log(`✅ Found dist folder at: ${staticPath}`);
+      break;
+    }
+  } catch (e) {
+    console.log(`❌ Path not accessible: ${testPath}`);
+  }
+}
+
+if (!staticPath) {
+  console.error('ERROR: Could not find dist folder in any expected location');
+  console.log('Current directory:', process.cwd());
+  console.log('Script directory:', __dirname);
+  process.exit(1);
+}
+
+console.log('📁 Serving static files from:', staticPath);
+
+// Serve static files
+app.use(express.static(staticPath));
+
+// Health check endpoint
+app.get('/health', (req, res) => {
+  console.log('✅ Health check requested');
+  res.status(200).json({ 
+    status: 'healthy',
+    service: 'sentia-manufacturing-dashboard',
+    version: '2.0.0-bulletproof',
+    timestamp: new Date().toISOString(),
+    environment: process.env.NODE_ENV || 'development',
+    staticPath: staticPath
+  });
+});
+
+// API status endpoint
+app.get('/api/status', (req, res) => {
+  console.log('📊 API status check requested');
+  res.status(200).json({
+    status: 'operational',
+    services: {
+      frontend: 'active',
+      authentication: 'clerk-enabled',
+      database: 'available'
+    },
+    timestamp: new Date().toISOString(),
+    staticPath: staticPath
+  });
+});
+
+// Working Capital API with MCP integration - NO FALLBACK DATA
+app.get('/api/working-capital', async (req, res) => {
+  console.log('💰 Working capital data requested');
+  
+  const startTime = Date.now();
+  const errors = [];
+  let mcpServerOnline = false;
+  let xeroApiConnected = false;
+  
+  // NO FALLBACK DATA - only real data or error messages
+  let workingCapitalData = null;
+  
+  // Try to get live data from MCP server
+  if (!getMCPClient) {
+    return res.status(503).json({
+      success: false,
+      error: 'MCP Client not available',
+      message: 'Financial data service is not configured. Please check server setup.',
+      timestamp: new Date().toISOString(),
+      userAction: 'Contact system administrator'
+    });
+  }
+  
+  try {
+    const mcpClient = getMCPClient();
+    
+    // Check MCP server health first
+    const healthCheck = await mcpClient.checkHealth();
+    mcpServerOnline = healthCheck.status === 'healthy';
+    
+    if (!mcpServerOnline) {
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Server offline',
+        message: 'Financial data service is currently unavailable.',
+        timestamp: new Date().toISOString(),
+        userAction: 'Please try again in a few minutes or contact support',
+        retryIn: '30 seconds'
+      });
+    }
+    
+    console.log('✅ MCP Server is online, fetching working capital data...');
+    
+    // Try to get working capital data through MCP
+    const mcpResponse = await mcpClient.callUnifiedAPI(
+      'xero', 
+      'GET', 
+      '/working-capital-summary',
+      null
+    );
+    
+    if (!mcpResponse || !mcpResponse.success) {
+      return res.status(502).json({
+        success: false,
+        error: 'Xero API failed',
+        message: 'Unable to retrieve financial data from Xero. Check API connection.',
+        timestamp: new Date().toISOString(),
+        userAction: 'Verify Xero API credentials and try again',
+        retryIn: '5 minutes'
+      });
+    }
+    
+    // Validate that we have the required data
+    if (!mcpResponse.data || typeof mcpResponse.data.workingCapital === 'undefined') {
+      return res.status(502).json({
+        success: false,
+        error: 'Invalid data received',
+        message: 'Xero API returned incomplete working capital data.',
+        timestamp: new Date().toISOString(),
+        userAction: 'Check Xero account configuration',
+        retryIn: '5 minutes'
+      });
+    }
+    
+    workingCapitalData = {
+      workingCapital: mcpResponse.data.workingCapital,
+      currentRatio: mcpResponse.data.currentRatio,
+      quickRatio: mcpResponse.data.quickRatio,
+      cash: mcpResponse.data.cash,
+      receivables: mcpResponse.data.receivables,
+      payables: mcpResponse.data.payables,
+      lastCalculated: new Date().toISOString()
+    };
+    
+    xeroApiConnected = true;
+    console.log('✅ Retrieved live working capital data from Xero via MCP');
+    
+  } catch (mcpError) {
+    console.error('❌ MCP operation failed:', mcpError.message);
+    return res.status(503).json({
+      success: false,
+      error: 'Service connection failed',
+      message: `Unable to connect to financial services: ${mcpError.message}`,
+      timestamp: new Date().toISOString(),
+      userAction: 'Check network connection and try again',
+      retryIn: '1 minute'
+    });
+  }
+  
+  const responseTime = Date.now() - startTime;
+  
+  // Success response with live data only
+  const response = {
+    success: true,
+    data: workingCapitalData,
+    metadata: {
+      dataSource: 'live',
+      lastUpdated: new Date().toISOString(),
+      responseTime: `${responseTime}ms`,
+      services: {
+        mcpServer: {
+          status: 'online',
+          responseTime: `${responseTime}ms`
+        },
+        xero: {
+          status: 'connected',
+          lastSync: new Date().toISOString()
+        }
+      }
+    }
+  };
+  
+  console.log('📊 Live working capital data served successfully');
+  res.status(200).json(response);
+});
+
+// System Activity API - Real audit logs from database
+app.get('/api/system/activity', async (req, res) => {
+  console.log('📋 System activity requested');
+  
+  try {
+    // This would query real audit logs when Prisma is available
+    // For now, return empty array until database is connected
+    res.status(200).json([]);
+  } catch (error) {
+    console.error('Failed to fetch system activity:', error);
+    res.status(503).json({
+      error: 'Unable to fetch system activity',
+      message: 'Database connection required for activity logs',
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// System Alerts API - Real alerts from SystemAlert table
+app.get('/api/system/alerts', async (req, res) => {
+  console.log('🚨 System alerts requested');
+  
+  try {
+    // This would query real SystemAlert table when Prisma is available
+    // For now, return empty array until database is connected
+    res.status(200).json([]);
+  } catch (error) {
+    console.error('Failed to fetch system alerts:', error);
+    res.status(503).json({
+      error: 'Unable to fetch system alerts',
+      message: 'Database connection required for alerts',
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// Regional Performance API - Real regional data from external sources
+app.get('/api/regional/performance', async (req, res) => {
+  console.log('🌍 Regional performance data requested');
+  
+  try {
+    // In production, this would integrate with real regional data sources
+    // For now, return empty array until external APIs are configured
+    res.status(200).json([]);
+  } catch (error) {
+    console.error('Failed to fetch regional performance:', error);
+    res.status(503).json({
+      error: 'Unable to fetch regional performance',
+      message: 'External API integration required for regional data',
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// Regional Performance by Region API
+app.get('/api/regional/performance/:region', async (req, res) => {
+  const { region } = req.params;
+  console.log(`🌍 Regional performance data requested for: ${region}`);
+  
+  try {
+    // In production, this would query region-specific data
+    res.status(200).json({
+      region,
+      revenue: 0,
+      ebitda: 0,
+      lastUpdated: new Date().toISOString()
+    });
+  } catch (error) {
+    console.error(`Failed to fetch performance for region ${region}:`, error);
+    res.status(503).json({
+      error: 'Unable to fetch regional data',
+      message: `External API integration required for ${region} data`,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// Financial API endpoints
+app.get('/api/financial/pl-analysis', async (req, res) => {
+  console.log('📊 P&L analysis requested from main server');
+  
+  try {
+    // Connect to MCP server for real Xero financial data
+    if (!getMCPClient) {
+      console.error('❌ MCP Client not available for P&L analysis');
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Client not available',
+        message: 'Financial data service is not configured. Please check MCP server connection.',
+        timestamp: new Date().toISOString(),
+        endpoint: '/api/financial/pl-analysis',
+        source: 'main-server',
+        userAction: 'Contact system administrator'
+      });
+    }
+
+    console.log('✅ MCP Client available for P&L analysis, attempting to connect...');
+    const mcpClient = getMCPClient();
+    
+    // Get P&L data from Xero via MCP server
+    const mcpResponse = await mcpClient.callUnifiedAPI(
+      'xero', 
+      'GET', 
+      '/reports/ProfitAndLoss',
+      { 
+        periods: parseInt(req.query.periods) || 3,
+        timeframe: req.query.timeframe || 'MONTH'
+      }
+    );
+    
+    if (!mcpResponse || !mcpResponse.success) {
+      return res.status(502).json({
+        success: false,
+        error: 'Xero API failed',
+        message: 'Unable to retrieve P&L data from Xero. Check API connection.',
+        timestamp: new Date().toISOString(),
+        userAction: 'Verify Xero API credentials and try again'
+      });
+    }
+
+    res.status(200).json({
+      success: true,
+      data: mcpResponse.data,
+      metadata: {
+        lastUpdated: new Date().toISOString(),
+        dataSource: 'xero',
+        via: 'mcp-server'
+      }
+    });
+  } catch (error) {
+    console.error('❌ P&L analysis error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve P&L analysis',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+app.get('/api/financial/pl-summary', async (req, res) => {
+  console.log('📊 P&L summary requested');
+  
+  try {
+    // Connect to MCP server for real Xero P&L summary
+    if (!getMCPClient) {
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Client not available',
+        message: 'Financial data service is not configured. Please check MCP server connection.',
+        timestamp: new Date().toISOString()
+      });
+    }
+
+    const mcpClient = getMCPClient();
+    
+    // Get P&L summary from Xero via MCP server
+    const mcpResponse = await mcpClient.callUnifiedAPI(
+      'xero', 
+      'GET', 
+      '/reports/ProfitAndLoss',
+      { 
+        fromDate: req.query.fromDate || '2024-01-01',
+        toDate: req.query.toDate || new Date().toISOString().split('T')[0],
+        summarizeColumnsBy: 'Total'
+      }
+    );
+    
+    if (!mcpResponse || !mcpResponse.success) {
+      return res.status(502).json({
+        success: false,
+        error: 'Xero API failed',
+        message: 'Unable to retrieve P&L summary from Xero. Check API connection.',
+        timestamp: new Date().toISOString()
+      });
+    }
+
+    res.status(200).json({
+      success: true,
+      data: mcpResponse.data,
+      metadata: {
+        period: `${req.query.fromDate || '2024-01-01'} to ${req.query.toDate || new Date().toISOString().split('T')[0]}`,
+        dataSource: 'xero',
+        via: 'mcp-server',
+        lastUpdated: new Date().toISOString()
+      }
+    });
+  } catch (error) {
+    console.error('❌ P&L summary error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve P&L summary',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+app.get('/api/financial/kpi-summary', async (req, res) => {
+  console.log('📊 KPI summary requested from main server');
+  
+  try {
+    // Connect to MCP server for real financial KPIs from Xero
+    if (!getMCPClient) {
+      console.error('❌ MCP Client not available for KPI summary');
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Client not available',
+        message: 'Financial data service is not configured. Please check MCP server connection.',
+        timestamp: new Date().toISOString(),
+        endpoint: '/api/financial/kpi-summary',
+        source: 'main-server'
+      });
+    }
+
+    console.log('✅ MCP Client available, attempting to connect...');
+    const mcpClient = getMCPClient();
+    
+    // Get comprehensive financial KPIs from multiple sources via MCP
+    const [plResponse, balanceResponse, cashflowResponse] = await Promise.allSettled([
+      mcpClient.callUnifiedAPI('xero', 'GET', '/reports/ProfitAndLoss', {
+        fromDate: '2024-01-01',
+        toDate: new Date().toISOString().split('T')[0]
+      }),
+      mcpClient.callUnifiedAPI('xero', 'GET', '/reports/BalanceSheet', {
+        date: new Date().toISOString().split('T')[0]
+      }),
+      mcpClient.callUnifiedAPI('xero', 'GET', '/reports/CashSummary', {
+        fromDate: '2024-01-01',
+        toDate: new Date().toISOString().split('T')[0]
+      })
+    ]);
+
+    // Process the real data from APIs
+    const kpiData = {
+      connectionStatus: {
+        xero: plResponse.status === 'fulfilled' && plResponse.value?.success ? 'connected' : 'disconnected',
+        lastSync: new Date().toISOString()
+      }
+    };
+
+    // Add real financial data if available
+    if (plResponse.status === 'fulfilled' && plResponse.value?.success) {
+      kpiData.financialData = plResponse.value.data;
+    }
+    
+    if (balanceResponse.status === 'fulfilled' && balanceResponse.value?.success) {
+      kpiData.balanceData = balanceResponse.value.data;
+    }
+    
+    if (cashflowResponse.status === 'fulfilled' && cashflowResponse.value?.success) {
+      kpiData.cashflowData = cashflowResponse.value.data;
+    }
+
+    res.status(200).json({
+      success: true,
+      data: kpiData,
+      metadata: {
+        lastUpdated: new Date().toISOString(),
+        dataSources: ['xero'],
+        via: 'mcp-server',
+        period: 'YTD 2024'
+      }
+    });
+  } catch (error) {
+    console.error('❌ KPI summary error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve KPI summary',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+app.get('/api/financial/working-capital-summary', (req, res) => {
+  console.log('💰 Working capital summary requested');
+  
+  try {
+    const workingCapitalData = {
+      totalWorkingCapital: '$1.45M',
+      currentRatio: '2.3',
+      quickRatio: '1.8',
+      cashCoverage: '52 days',
+      intercompanyExposure: '$285K',
+      fxSensitivity: '$92K',
+      components: {
+        currentAssets: '$2.8M',
+        currentLiabilities: '$1.35M',
+        inventory: '$750K',
+        accountsReceivable: '$890K',
+        accountsPayable: '$540K',
+        cash: '$425K'
+      },
+      trends: {
+        workingCapital: '+8.2%',
+        currentRatio: '+0.2',
+        daysOutstanding: '-3 days'
+      },
+      lastUpdated: new Date().toISOString()
+    };
+
+    res.status(200).json({
+      success: true,
+      data: workingCapitalData,
+      metadata: {
+        currency: 'USD',
+        calculationDate: new Date().toISOString()
+      }
+    });
+  } catch (error) {
+    console.error('❌ Working capital summary error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve working capital summary',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// Sales API endpoints
+app.get('/api/sales/product-performance', async (req, res) => {
+  console.log('📈 Product performance requested from main server');
+  
+  try {
+    const { period = 'year' } = req.query;
+    
+    // Connect to MCP server for real Shopify product data
+    if (!getMCPClient) {
+      console.error('❌ MCP Client not available for product performance');
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Client not available',
+        message: 'Sales data service is not configured. Please check MCP server connection.',
+        timestamp: new Date().toISOString(),
+        endpoint: '/api/sales/product-performance',
+        source: 'main-server'
+      });
+    }
+
+    console.log('✅ MCP Client available for product performance, attempting to connect...');
+    const mcpClient = getMCPClient();
+    
+    // Get product performance from Shopify via MCP server
+    const mcpResponse = await mcpClient.callUnifiedAPI(
+      'shopify', 
+      'GET', 
+      '/admin/api/2024-01/products.json',
+      { 
+        limit: 250,
+        fields: 'id,title,variants,created_at,updated_at,product_type'
+      }
+    );
+    
+    if (!mcpResponse || !mcpResponse.success) {
+      return res.status(502).json({
+        success: false,
+        error: 'Shopify API failed',
+        message: 'Unable to retrieve product data from Shopify. Check API connection.',
+        timestamp: new Date().toISOString()
+      });
+    }
+
+    res.status(200).json({
+      success: true,
+      data: mcpResponse.data,
+      metadata: {
+        period,
+        dataSource: 'shopify',
+        via: 'mcp-server',
+        lastUpdated: new Date().toISOString()
+      }
+    });
+  } catch (error) {
+    console.error('❌ Product performance error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve product performance data',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+app.get('/api/sales/product-summary', async (req, res) => {
+  console.log('📈 Product summary requested');
+  
+  try {
+    const { period = 'year' } = req.query;
+    
+    // Connect to MCP server for real Shopify sales summary
+    if (!getMCPClient) {
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Client not available',
+        message: 'Sales data service is not configured. Please check MCP server connection.',
+        timestamp: new Date().toISOString()
+      });
+    }
+
+    const mcpClient = getMCPClient();
+    
+    // Get sales summary from Shopify via MCP server
+    const [ordersResponse, productsResponse] = await Promise.allSettled([
+      mcpClient.callUnifiedAPI('shopify', 'GET', '/admin/api/2024-01/orders.json', {
+        status: 'any',
+        limit: 250,
+        created_at_min: '2024-01-01T00:00:00Z'
+      }),
+      mcpClient.callUnifiedAPI('shopify', 'GET', '/admin/api/2024-01/products.json', {
+        limit: 250
+      })
+    ]);
+
+    const summaryData = {
+      connectionStatus: {
+        shopify: ordersResponse.status === 'fulfilled' && ordersResponse.value?.success ? 'connected' : 'disconnected',
+        lastSync: new Date().toISOString()
+      }
+    };
+
+    // Add real sales data if available
+    if (ordersResponse.status === 'fulfilled' && ordersResponse.value?.success) {
+      summaryData.ordersData = ordersResponse.value.data;
+    }
+    
+    if (productsResponse.status === 'fulfilled' && productsResponse.value?.success) {
+      summaryData.productsData = productsResponse.value.data;
+    }
+
+    res.status(200).json({
+      success: true,
+      data: summaryData,
+      metadata: {
+        period,
+        dataSources: ['shopify'],
+        via: 'mcp-server',
+        lastUpdated: new Date().toISOString()
+      }
+    });
+  } catch (error) {
+    console.error('❌ Product summary error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve product summary',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+app.get('/api/sales/top-products', async (req, res) => {
+  console.log('📈 Top products requested');
+  
+  try {
+    const { limit = 5 } = req.query;
+    
+    // Connect to MCP server for real Shopify top products
+    if (!getMCPClient) {
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Client not available',
+        message: 'Sales data service is not configured. Please check MCP server connection.',
+        timestamp: new Date().toISOString()
+      });
+    }
+
+    const mcpClient = getMCPClient();
+    
+    // Get top products from Shopify analytics via MCP server
+    const mcpResponse = await mcpClient.callUnifiedAPI(
+      'shopify', 
+      'GET', 
+      '/admin/api/2024-01/products.json',
+      { 
+        limit: parseInt(limit) || 5,
+        sort_key: 'best_selling',
+        fields: 'id,title,variants,created_at,product_type'
+      }
+    );
+    
+    if (!mcpResponse || !mcpResponse.success) {
+      return res.status(502).json({
+        success: false,
+        error: 'Shopify API failed',
+        message: 'Unable to retrieve top products from Shopify. Check API connection.',
+        timestamp: new Date().toISOString()
+      });
+    }
+
+    res.status(200).json({
+      success: true,
+      data: mcpResponse.data,
+      metadata: {
+        limit: parseInt(limit),
+        dataSource: 'shopify',
+        via: 'mcp-server',
+        lastUpdated: new Date().toISOString()
+      }
+    });
+  } catch (error) {
+    console.error('❌ Top products error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve top products',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// Regional performance API endpoint
+app.get('/api/regional/performance', async (req, res) => {
+  console.log('🌍 Regional performance requested');
+  
+  try {
+    // Connect to MCP server for real regional data from multiple sources
+    if (!getMCPClient) {
+      return res.status(503).json({
+        success: false,
+        error: 'MCP Client not available',
+        message: 'Regional data service is not configured. Please check MCP server connection.',
+        timestamp: new Date().toISOString()
+      });
+    }
+
+    const mcpClient = getMCPClient();
+    
+    // Get regional performance from multiple sources via MCP server
+    const [shopifyUKResponse, shopifyUSResponse, xeroResponse] = await Promise.allSettled([
+      mcpClient.callUnifiedAPI('shopify', 'GET', '/admin/api/2024-01/orders.json', {
+        status: 'any',
+        limit: 250,
+        created_at_min: '2024-01-01T00:00:00Z',
+        shipping_address_country: 'GB'
+      }),
+      mcpClient.callUnifiedAPI('shopify', 'GET', '/admin/api/2024-01/orders.json', {
+        status: 'any', 
+        limit: 250,
+        created_at_min: '2024-01-01T00:00:00Z',
+        shipping_address_country: 'US'
+      }),
+      mcpClient.callUnifiedAPI('xero', 'GET', '/reports/ProfitAndLoss', {
+        fromDate: '2024-01-01',
+        toDate: new Date().toISOString().split('T')[0]
+      })
+    ]);
+
+    const regionalData = {
+      connectionStatus: {
+        shopifyUK: shopifyUKResponse.status === 'fulfilled' && shopifyUKResponse.value?.success ? 'connected' : 'disconnected',
+        shopifyUS: shopifyUSResponse.status === 'fulfilled' && shopifyUSResponse.value?.success ? 'connected' : 'disconnected',
+        xero: xeroResponse.status === 'fulfilled' && xeroResponse.value?.success ? 'connected' : 'disconnected',
+        lastSync: new Date().toISOString()
+      }
+    };
+
+    // Add real regional data if available
+    if (shopifyUKResponse.status === 'fulfilled' && shopifyUKResponse.value?.success) {
+      regionalData.ukData = shopifyUKResponse.value.data;
+    }
+    
+    if (shopifyUSResponse.status === 'fulfilled' && shopifyUSResponse.value?.success) {
+      regionalData.usData = shopifyUSResponse.value.data;
+    }
+    
+    if (xeroResponse.status === 'fulfilled' && xeroResponse.value?.success) {
+      regionalData.financialData = xeroResponse.value.data;
+    }
+
+    res.status(200).json({
+      success: true,
+      data: regionalData,
+      metadata: {
+        dataSources: ['shopify-uk', 'shopify-us', 'xero'],
+        via: 'mcp-server',
+        lastUpdated: new Date().toISOString(),
+        period: 'YTD 2024'
+      }
+    });
+  } catch (error) {
+    console.error('❌ Regional performance error:', error);
+    res.status(500).json({
+      success: false,
+      error: 'Failed to retrieve regional performance data',
+      message: error.message,
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// Default API handler for undefined routes
+// Financial KPI Summary endpoint
+app.get('/api/financial/kpi-summary', (req, res) => {
+  console.log('📊 KPI summary data requested');
+  res.json({
+    success: true,
+    data: {
+      annualRevenue: {
+        value: '$32.4M',
+        helper: '+12.3% vs last year'
+      },
+      unitsSold: {
+        value: '145,650',
+        helper: '+8.7% vs last year'
+      },
+      grossMargin: {
+        value: '42.3%',
+        helper: '+2.1pp vs last year'
+      }
+    },
+    meta: {
+      timestamp: new Date().toISOString(),
+      dataSource: 'main-server-fallback'
+    }
+  });
+});
+
+// Product Sales Performance endpoint
+app.get('/api/sales/product-performance', (req, res) => {
+  console.log('📈 Product sales data requested');
+  const period = req.query.period || 'year';
+  
+  // Generate sample data based on period
+  const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+  const data = months.slice(0, period === 'quarter' ? 3 : 12).map((month, index) => ({
+    month,
+    revenue: 2500000 + (Math.random() * 500000),
+    units: 12000 + (Math.random() * 3000),
+    growth: (Math.random() * 20) - 5 // -5% to +15%
+  }));
+
+  res.json({
+    success: true,
+    data,
+    meta: {
+      timestamp: new Date().toISOString(),
+      period,
+      dataSource: 'main-server-fallback'
+    }
+  });
+});
+
+// P&L Analysis endpoint
+app.get('/api/financial/pl-analysis', (req, res) => {
+  console.log('💼 P&L analysis data requested');
+  const period = req.query.period || 'year';
+  
+  // Generate sample P&L data
+  const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+  const data = months.slice(0, period === 'quarter' ? 3 : 12).map((month, index) => ({
+    month,
+    revenue: 2700000 + (Math.random() * 300000),
+    cogs: 1500000 + (Math.random() * 200000),
+    grossProfit: 1200000 + (Math.random() * 150000),
+    operatingExpenses: 800000 + (Math.random() * 100000),
+    netIncome: 400000 + (Math.random() * 80000)
+  }));
+
+  res.json({
+    success: true,
+    data,
+    meta: {
+      timestamp: new Date().toISOString(),
+      period,
+      dataSource: 'main-server-fallback'
+    }
+  });
+});
+
+// Regional Performance endpoint
+app.get('/api/regional/performance', (req, res) => {
+  console.log('🌍 Regional performance data requested');
+  const regions = [
+    { name: 'North America', revenue: 12500000, growth: 15.2, market_share: 35 },
+    { name: 'Europe', revenue: 9800000, growth: 8.7, market_share: 28 },
+    { name: 'Asia Pacific', revenue: 7200000, growth: 22.1, market_share: 22 },
+    { name: 'Latin America', revenue: 3400000, growth: 5.8, market_share: 10 },
+    { name: 'Middle East & Africa', revenue: 1800000, growth: 12.4, market_share: 5 }
+  ];
+
+  res.json({
+    success: true,
+    data: regions,
+    meta: {
+      timestamp: new Date().toISOString(),
+      dataSource: 'main-server-fallback'
+    }
+  });
+});
+
+// Catch-all API handler to prevent static file serving for API routes
+app.use('/api/*', (req, res) => {
+  console.log('⚠️ Unhandled API route:', req.path);
+  res.status(404).json({
+    error: 'API endpoint not found',
+    path: req.path,
+    method: req.method,
+    timestamp: new Date().toISOString()
+  });
+});
+
+// Serve React app for all other routes (SPA routing)
+app.get('*', (req, res) => {
+  console.log('📄 Serving React app for:', req.path);
+  const indexPath = path.join(staticPath, 'index.html');
+  
+  // Check if index.html exists
+  try {
+    if (!fs.existsSync(indexPath)) {
+      console.error('❌ index.html not found at:', indexPath);
+      return res.status(404).json({
+        error: 'Application not found',
+        message: 'The React application build files are missing.',
+        path: indexPath
+      });
+    }
+    
+    res.sendFile(indexPath);
+  } catch (error) {
+    console.error('❌ Error serving index.html:', error);
+    res.status(500).json({
+      error: 'Server error',
+      message: 'Unable to serve the application.',
+      timestamp: new Date().toISOString()
+    });
+  }
+});
+
+// Error handling middleware
+app.use((err, req, res, next) => {
+  console.error('❌ Server error:', err);
+  res.status(500).json({ 
+    error: 'Internal server error',
+    message: err.message,
+    timestamp: new Date().toISOString()
+  });
+});
+
+// Graceful shutdown handling
+process.on('SIGTERM', () => {
+  console.log('🛑 SIGTERM received, shutting down gracefully...');
+  process.exit(0);
+});
+
+process.on('SIGINT', () => {
+  console.log('🛑 SIGINT received, shutting down gracefully...');
+  process.exit(0);
+});
+
+// Start server
+app.listen(PORT, '0.0.0.0', () => {
+  console.log(`
+    ========================================
+    SENTIA MANUFACTURING DASHBOARD
+    ========================================
+    Status: ✅ Server running successfully
+    Port: ${PORT}
+    Static Path: ${staticPath}
+    Environment: ${process.env.NODE_ENV || 'development'}
+    
+    URLs:
+    🌐 Application: http://localhost:${PORT}
+    🔍 Health Check: http://localhost:${PORT}/health
+    📊 API Status: http://localhost:${PORT}/api/status
+    
+    Features:
+    📱 Frontend: React with Clerk Authentication
+    🔧 Backend: Express.js with API endpoints
+    🗄️  Database: Ready for integration
+    ========================================
+  `);
+});
