+++ package.json
@@ -13,7 +13,6 @@
     "build:dev": "vite build --mode development",
     "build:test": "vite build --mode testing",
     "start": "node server.js",
-    "start:render": "node render-start.js",
     "preview": "vite preview",
     "test": "echo 'No tests configured' && exit 0",
     "test:run": "vitest run",
