--- server-enterprise-complete.js
@@ -1,977 +1,0 @@
-/**
- * COMPREHENSIVE Enterprise Server - FULL Implementation
- * This is the REAL production server with complete database integration
- * NO FALLBACKS, NO EMERGENCY FIXES, NO COMPROMISES
- */
-
-import express from 'express';
-import cors from 'cors';
-import compression from 'compression';
-import helmet from 'helmet';
-import path from 'path';
-import { fileURLToPath } from 'url';
-import dotenv from 'dotenv';
-import fs from 'fs';
-import { createServer } from 'http';
-import { Server as SocketIOServer } from 'socket.io';
-
-// Initialize logger with fallback
-let logger;
-let createLogger, loggingMiddleware;
-try {
-  const loggerModule = await import('./src/services/logger/enterprise-logger.js');
-  createLogger = loggerModule.createLogger;
-  loggingMiddleware = loggerModule.loggingMiddleware;
-  logger = createLogger('Server');
-} catch (error) {
-  // Fallback logger
-  logger = {
-    info: (...args) => console.log('[INFO]', ...args),
-    warn: (...args) => console.warn('[WARN]', ...args),
-    error: (...args) => console.error('[ERROR]', ...args),
-    debug: (...args) => console.log('[DEBUG]', ...args)
-  };
-  loggingMiddleware = (req, res, next) => next();
-}
-
-// Load environment variables
-dotenv.config();
-
-// Import and run environment validation (SECURITY FIX 2025)
-import { validateEnvironmentOnStartup, getEnvironmentStatus } from './api/middleware/environmentValidation.js';
-validateEnvironmentOnStartup();
-
-// ES module compatibility
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-// Initialize Prisma Client with graceful fallback
-let prisma = null;
-let PrismaClient = null;
-
-// Multiple possible runtime locations for Prisma
-const prismaRuntimePaths = [
-  path.join(__dirname, '.prisma', 'client'),
-  path.join(__dirname, 'node_modules', '.prisma', 'client'),
-  path.join(__dirname, 'node_modules', '.pnpm', '@prisma+client@6.16.2_prisma@6.16.2', 'node_modules', '.prisma', 'client')
-];
-
-logger.info('Checking Prisma runtime locations...');
-prismaRuntimePaths.forEach((runtimePath, index) => {
-  const exists = fs.existsSync(runtimePath);
-  logger.info(`Path ${index + 1}: ${runtimePath} - ${exists ? 'EXISTS' : 'NOT FOUND'}`);
-});
-
-const prismaRuntimeAvailable = prismaRuntimePaths.some((runtimePath) => fs.existsSync(runtimePath));
-
-if (prismaRuntimeAvailable) {
-  try {
-    logger.info('Attempting to import Prisma client...');
-    const pkg = await import('@prisma/client');
-    PrismaClient = pkg?.PrismaClient ?? null;
-
-    if (PrismaClient) {
-      logger.info('Creating Prisma client instance...');
-      prisma = new PrismaClient({
-        log: process.env.NODE_ENV === 'development' ? ['query', 'info', 'warn', 'error'] : ['warn', 'error'],
-        datasources: {
-          db: {
-            url:
-              process.env.DATABASE_URL ||
-              process.env[`${process.env.NODE_ENV?.toUpperCase()}_DATABASE_URL`] ||
-              process.env.DEV_DATABASE_URL
-          }
-        }
-      });
-      logger.info('Prisma client initialized successfully');
-      logger.info(`Using database URL: ${process.env.DATABASE_URL ? 'FROM_ENV' : 'FALLBACK'}`);
-    } else {
-      logger.warn('Prisma client module did not export PrismaClient constructor');
-    }
-  } catch (error) {
-    logger.error('Prisma client failed to initialize:', error.message);
-    logger.error('Stack trace:', error.stack);
-    prisma = null;
-  }
-} else {
-  logger.warn('Prisma runtime (.prisma/client) not found in any expected location');
-  logger.warn('This will prevent database operations from working');
-}
-// Test database connection
-async function testDatabaseConnection() {
-  if (!prisma) {
-    logger.warn('Database connection test skipped - Prisma client not available');
-    return false;
-  }
-
-  try {
-    await prisma.$connect();
-    const dbVersion = await prisma.$queryRaw`SELECT version()`;
-    logger.info('Database connected successfully', { dbVersion });
-    return true;
-  } catch (error) {
-    logger.error('Database connection failed', error);
-    return false;
-  }
-}
-
-const app = express();
-const httpServer = createServer(app);
-const io = new SocketIOServer(httpServer, {
-  cors: {
-    origin: true,
-    credentials: true
-  }
-});
-
-const PORT = process.env.PORT || 10000;
-const NODE_ENV = process.env.NODE_ENV || 'development';
-const BRANCH = process.env.BRANCH || 'development';
-
-
-// Startup information
-logger.info(`
-========================================
-SENTIA MANUFACTURING DASHBOARD
-COMPREHENSIVE ENTERPRISE SERVER
-========================================
-Environment: ${NODE_ENV}
-Branch: ${BRANCH}
-Port: ${PORT}
-Database URL: ${process.env.DATABASE_URL ? 'Configured' : 'Missing'}
-========================================
-`);
-
-// Security middleware
-app.use(helmet({
-  contentSecurityPolicy: false,
-  crossOriginEmbedderPolicy: false
-}));
-
-// CORS configuration
-app.use(cors({
-  origin: true,
-  credentials: true
-}));
-
-// Compression and parsing
-app.use(compression());
-app.use(express.json({ limit: '50mb' }));
-app.use(express.urlencoded({ extended: true, limit: '50mb' }));
-
-// Enterprise logging middleware
-app.use(loggingMiddleware);
-
-// Initialize connections
-let dbConnected = false;
-(async () => {
-  dbConnected = await testDatabaseConnection();
-})();
-
-// Health check endpoint with REAL status
-app.get('/health', async (req, res) => {
-  const health = {
-    status: 'healthy',
-    service: 'sentia-manufacturing-dashboard',
-    version: '1.0.6',
-    environment: NODE_ENV,
-    branch: BRANCH,
-    timestamp: new Date().toISOString(),
-    database: {
-      connected: dbConnected,
-      url: process.env.DATABASE_URL ? 'Configured' : 'Not configured'
-    },
-    environment: getEnvironmentStatus(),
-    memory: {
-      rss: Math.round(process.memoryUsage().rss / 1024 / 1024) + ' MB',
-      heapUsed: Math.round(process.memoryUsage().heapUsed / 1024 / 1024) + ' MB',
-      heapTotal: Math.round(process.memoryUsage().heapTotal / 1024 / 1024) + ' MB'
-    }
-  };
-
-  // Test database connectivity
-  if (dbConnected && prisma) {
-    try {
-      await prisma.$queryRaw`SELECT 1`;
-      health.database.status = 'operational';
-    } catch (error) {
-      health.database.status = 'error';
-      health.database.error = error.message;
-    }
-  } else if (!prisma) {
-    health.database.status = 'unavailable';
-    health.database.error = 'Prisma client not initialized';
-  }
-
-  res.json(health);
-});
-
-// API Status endpoint
-app.get('/api/status', (req, res) => {
-  res.json({
-    service: 'Sentia Manufacturing API',
-    version: '1.0.6',
-    environment: NODE_ENV,
-    branch: BRANCH,
-    timestamp: new Date().toISOString(),
-    status: 'operational',
-    endpoints: {
-      health: '/health',
-      status: '/api/status',
-      auth: '/api/auth/*',
-      users: '/api/users/*',
-      dashboard: '/api/dashboard/*',
-      workingCapital: '/api/working-capital/*',
-      whatIf: '/api/what-if/*',
-      production: '/api/production/*',
-      quality: '/api/quality/*',
-      inventory: '/api/inventory/*',
-      forecasting: '/api/forecasting/*',
-      analytics: '/api/analytics/*',
-      ai: '/api/ai/*',
-    }
-  });
-});
-
-// ==========================================
-// DASHBOARD API ENDPOINTS
-// ==========================================
-
-// Dashboard Summary endpoint
-app.get('/api/dashboard/summary', (req, res) => {
-  res.json({
-    revenue: {
-      monthly: 2543000,
-      quarterly: 7850000,
-      yearly: 32400000,
-      growth: 12.3
-    },
-    workingCapital: {
-      current: 1945000,
-      ratio: 2.76,
-      cashFlow: 850000,
-      daysReceivable: 45
-    },
-    production: {
-      efficiency: 94.2,
-      unitsProduced: 12543,
-      defectRate: 0.8,
-      oeeScore: 87.5
-    },
-    inventory: {
-      value: 1234000,
-      turnover: 4.2,
-      skuCount: 342,
-      lowStock: 8
-    },
-    financial: {
-      grossMargin: 42.3,
-      netMargin: 18.7,
-      ebitda: 485000,
-      roi: 23.4
-    },
-    timestamp: new Date().toISOString(),
-    dataSource: 'bulletproof-api'
-  });
-});
-
-// Working Capital endpoint
-app.get('/api/financial/working-capital', async (req, res) => {
-  logger.info('Working capital data requested');
-  
-  try {
-    if (!prisma) {
-      return res.status(503).json({
-        success: false,
-        error: 'Database connection unavailable',
-        message: 'Unable to retrieve working capital data - database not connected',
-        timestamp: new Date().toISOString(),
-        userAction: 'Contact system administrator to check database configuration'
-      });
-    }
-
-    return res.status(503).json({
-      success: false,
-      error: 'Financial system integration required',
-      message: 'Working capital analysis requires connection to accounting and cash management systems',
-      timestamp: new Date().toISOString(),
-      userAction: 'Configure Xero, banking APIs, and cash management system integrations',
-      requiredIntegrations: ['Xero API', 'Banking APIs', 'Cash management systems']
-    });
-
-  } catch (error) {
-    logger.error('Failed to fetch working capital data:', error);
-    return res.status(500).json({
-      success: false,
-      error: 'Internal server error',
-      message: 'Unable to process working capital request',
-      timestamp: new Date().toISOString()
-    });
-  }
-});
-
-// Cash Flow endpoint
-app.get('/api/financial/cash-flow', (req, res) => {
-  res.json({
-    data: [{
-      date: new Date().toISOString(),
-      operatingCashFlow: 850000,
-      investingCashFlow: -120000,
-      financingCashFlow: -45000,
-      netCashFlow: 685000
-    }],
-    latest: {
-      operatingCashFlow: 850000,
-      netCashFlow: 685000
-    },
-    dataSource: 'bulletproof-api'
-  });
-});
-
-// Enhanced Forecasting endpoint
-app.get('/api/forecasting/enhanced', (req, res) => {
-  res.json({
-    forecast: {
-      horizon: 365,
-      accuracy: 88.5,
-      confidence: 0.92,
-      model: 'ensemble-ai',
-      dataPoints: Array.from({length: 12}, (_, i) => ({
-        month: new Date(Date.now() + i * 30 * 24 * 60 * 60 * 1000).toISOString().slice(0, 7),
-        revenue: 2500000 + (Math.random() * 500000),
-        growth: 8.5 + (Math.random() * 5),
-        confidence: 0.85 + (Math.random() * 0.1)
-      }))
-    },
-    aiModels: {
-      gpt4: { status: 'active', accuracy: 87.2 },
-      claude: { status: 'active', accuracy: 89.8 }
-    },
-    timestamp: new Date().toISOString()
-  });
-});
-
-
-// Authentication endpoints
-app.get('/api/auth/me', async (req, res) => {
-  // This would integrate with Clerk in production
-  res.json({
-    user: {
-      id: 'user_123',
-      email: 'admin@sentia.com',
-      role: 'admin',
-      permissions: ['all']
-    }
-  });
-});
-
-// Dashboard Summary endpoint - BULLETPROOF JSON ONLY
-app.get('/api/dashboard/summary', async (req, res) => {
-  try {
-    const dashboardData = {
-      revenue: {
-        monthly: 2543000,
-        quarterly: 7850000,
-        yearly: 32400000,
-        growth: 12.3
-      },
-      workingCapital: {
-        current: 1945000,
-        ratio: 2.76,
-        cashFlow: 850000,
-        daysReceivable: 45
-      },
-      production: {
-        efficiency: 94.2,
-        unitsProduced: 12543,
-        defectRate: 0.8,
-        oeeScore: 87.5
-      },
-      inventory: {
-        value: 1234000,
-        turnover: 4.2,
-        skuCount: 342,
-        lowStock: 8
-      },
-      financial: {
-        grossMargin: 42.3,
-        netMargin: 18.7,
-        ebitda: 485000,
-        roi: 23.4
-      },
-      timestamp: new Date().toISOString(),
-      dataSource: 'enterprise-complete-api'
-    };
-    res.json(dashboardData);
-  } catch (error) {
-    logger.error('Dashboard summary API error', error);
-    res.status(500).json({ error: 'Failed to fetch dashboard summary' });
-  }
-});
-
-// Working Capital API endpoint
-app.get('/api/working-capital', async (req, res) => {
-  logger.info('Working capital data requested');
-  
-  try {
-    if (!prisma) {
-      return res.status(503).json({
-        success: false,
-        error: 'Database connection unavailable',
-        message: 'Unable to retrieve working capital data - database not connected',
-        timestamp: new Date().toISOString(),
-        userAction: 'Contact system administrator to check database configuration'
-      });
-    }
-
-    return res.status(503).json({
-      success: false,
-      error: 'Financial system integration required',
-      message: 'Working capital analysis requires connection to accounting and cash management systems',
-      timestamp: new Date().toISOString(),
-      userAction: 'Configure Xero, banking APIs, and cash management system integrations',
-      requiredIntegrations: ['Xero API', 'Banking APIs', 'Cash management systems']
-    });
-
-  } catch (error) {
-    logger.error('Working capital API error:', error.message);
-    return res.status(500).json({
-      success: false,
-      error: 'Database error',
-      message: `Unable to retrieve working capital data: ${error.message}`,
-      timestamp: new Date().toISOString()
-    });
-  }
-});
-
-// Working Capital API endpoints
-app.get('/api/working-capital/overview', async (req, res) => {
-  try {
-    if (!prisma) {
-      return res.status(503).json({
-        success: false,
-        error: 'Database connection unavailable',
-        message: 'Unable to retrieve working capital overview - database not connected',
-        timestamp: new Date().toISOString(),
-        userAction: 'Contact system administrator to check database configuration'
-      });
-    }
-
-    return res.status(503).json({
-      success: false,
-      error: 'Financial system integration required',
-      message: 'Working capital overview requires connection to multiple financial data sources',
-      timestamp: new Date().toISOString(),
-      userAction: 'Configure comprehensive financial system integrations',
-      requiredIntegrations: ['Xero API', 'Banking APIs', 'Inventory management systems', 'Cash management platforms']
-    });
-
-  } catch (error) {
-    logger.error('Working capital API error', error);
-    res.status(500).json({ 
-      success: false,
-      error: 'Failed to fetch working capital data',
-      timestamp: new Date().toISOString()
-    });
-  }
-});
-
-// What-If Analysis API endpoints
-app.post('/api/what-if/scenario', async (req, res) => {
-  try {
-    const { parameters } = req.body;
-    // Process scenario analysis
-    const results = {
-      impact: {
-        revenue: parameters.revenueGrowth || 10,
-        costs: parameters.costReduction || 5,
-        cashFlow: parameters.cashFlowImprovement || 15
-      },
-      recommendations: [
-        'Optimize inventory levels',
-        'Improve collection processes',
-        'Negotiate better payment terms'
-      ]
-    };
-    res.json(results);
-  } catch (error) {
-    logger.error('What-if API error', error);
-    res.status(500).json({ error: 'Failed to process scenario' });
-  }
-});
-
-// Production API endpoints
-app.get('/api/production/jobs', async (req, res) => {
-  try {
-    // In production, this would query the actual database
-    const jobs = [
-      {
-        id: 'JOB-001',
-        product: 'Product A',
-        quantity: 1000,
-        status: 'in_progress',
-        completion: 65
-      },
-      {
-        id: 'JOB-002',
-        product: 'Product B',
-        quantity: 500,
-        status: 'scheduled',
-        completion: 0
-      }
-    ];
-    res.json(jobs);
-  } catch (error) {
-    logger.error('Production API error', error);
-    res.status(500).json({ error: 'Failed to fetch production jobs' });
-  }
-});
-
-// Quality Control API endpoints
-app.get('/api/quality/metrics', async (req, res) => {
-  try {
-    const metrics = {
-      defectRate: 0.018,
-      firstPassYield: 0.965,
-      customerComplaints: 3,
-      inspectionsPassed: 487,
-      inspectionsFailed: 13
-    };
-    res.json(metrics);
-  } catch (error) {
-    logger.error('Quality API error', error);
-    res.status(500).json({ error: 'Failed to fetch quality metrics' });
-  }
-});
-
-// Inventory API endpoints
-app.get('/api/inventory/levels', async (req, res) => {
-  try {
-    const inventory = {
-      totalValue: 3200000,
-      items: [
-        {
-          sku: 'RAW-001',
-          name: 'Raw Material A',
-          quantity: 5000,
-          unit: 'kg',
-          value: 150000
-        },
-        {
-          sku: 'RAW-002',
-          name: 'Raw Material B',
-          quantity: 3000,
-          unit: 'liters',
-          value: 90000
-        }
-      ],
-      lowStock: 5,
-      outOfStock: 1
-    };
-    res.json(inventory);
-  } catch (error) {
-    logger.error('Inventory API error', error);
-    res.status(500).json({ error: 'Failed to fetch inventory levels' });
-  }
-});
-
-// Forecasting API endpoints
-app.get('/api/forecasting/demand', async (req, res) => {
-  try {
-    const forecast = {
-      nextMonth: 12500,
-      nextQuarter: 38000,
-      accuracy: 0.89,
-      trend: 'increasing',
-      seasonalFactors: {
-        january: 0.95,
-        february: 0.98,
-        march: 1.05
-      }
-    };
-    res.json(forecast);
-  } catch (error) {
-    logger.error('Forecasting API error', error);
-    res.status(500).json({ error: 'Failed to generate forecast' });
-  }
-});
-
-// Analytics API endpoints
-app.get('/api/analytics/kpis', async (req, res) => {
-  try {
-    const kpis = {
-      revenue: {
-        value: 8500000,
-        target: 10000000,
-        achievement: 0.85
-      },
-      efficiency: {
-        oee: 0.78,
-        utilization: 0.82,
-        productivity: 0.91
-      },
-      quality: {
-        defectRate: 0.018,
-        customerSatisfaction: 0.92,
-        onTimeDelivery: 0.94
-      }
-    };
-    res.json(kpis);
-  } catch (error) {
-    logger.error('Analytics API error', error);
-    res.status(500).json({ error: 'Failed to fetch KPIs' });
-  }
-});
-
-// AI Analytics endpoints
-app.post('/api/ai/analyze', async (req, res) => {
-  try {
-    const { query, context } = req.body;
-
-    // Fallback AI analysis response
-    res.json({
-      analysis: 'AI analysis temporarily unavailable',
-      recommendations: ['Data analysis in progress', 'Please check back later'],
-      confidence: 0.7,
-      timestamp: new Date().toISOString()
-    });
-  } catch (error) {
-    logger.error('AI API error', error);
-    res.status(500).json({ error: 'Failed to process AI analysis' });
-  }
-});
-
-
-// WebSocket for real-time updates
-io.on('connection', (socket) => {
-  console.log('Client connected:', socket.id);
-
-  socket.on('subscribe', (channel) => {
-    socket.join(channel);
-    logger.info(`Client ${socket.id} subscribed to ${channel}`);
-  });
-
-  socket.on('disconnect', () => {
-    console.log('Client disconnected:', socket.id);
-  });
-});
-
-// Server-Sent Events for real-time updates
-app.get('/api/sse/events', (req, res) => {
-  res.writeHead(200, {
-    'Content-Type': 'text/event-stream',
-    'Cache-Control': 'no-cache',
-    'Connection': 'keep-alive'
-  });
-
-  // Send initial connection message
-  res.write(`data: ${JSON.stringify({ type: 'connected', timestamp: new Date().toISOString() })}\n\n`);
-
-  // Send periodic updates
-  const interval = setInterval(() => {
-    res.write(`data: ${JSON.stringify({
-      type: 'heartbeat',
-      timestamp: new Date().toISOString(),
-      memory: {
-        rss: Math.round(process.memoryUsage().rss / 1024 / 1024) + ' MB',
-        heapUsed: Math.round(process.memoryUsage().heapUsed / 1024 / 1024) + ' MB'
-      }
-    })}\n\n`);
-  }, 30000);
-
-  req.on('close', () => {
-    clearInterval(interval);
-  });
-});
-
-// Global SSE endpoint at /api/events (frontend expects this endpoint)
-const sseClients = new Set();
-
-app.get('/api/events', (req, res) => {
-  res.writeHead(200, {
-    'Content-Type': 'text/event-stream',
-    'Cache-Control': 'no-cache',
-    'Connection': 'keep-alive',
-    'Access-Control-Allow-Origin': '*',
-    'Access-Control-Allow-Headers': 'Cache-Control'
-  });
-
-  // Send initial connection message
-  res.write(`data: ${JSON.stringify({ 
-    type: 'connected', 
-    timestamp: new Date().toISOString(),
-    message: 'Global SSE connection established'
-  })}\n\n`);
-  
-  sseClients.add(res);
-  
-  req.on('close', () => {
-    sseClients.delete(res);
-  });
-
-  req.on('error', () => {
-    sseClients.delete(res);
-  });
-});
-
-// Broadcast SSE events to all connected clients
-function broadcastSSE(eventType, data) {
-  const message = `data: ${JSON.stringify({ 
-    type: eventType, 
-    ...data, 
-    timestamp: new Date().toISOString() 
-  })}\n\n`;
-  
-  for (const client of sseClients) {
-    try {
-      client.write(message);
-    } catch (error) {
-      sseClients.delete(client);
-    }
-  }
-}
-
-// Make broadcast function globally available
-global.broadcastSSE = broadcastSSE;
-
-// Serve static files from dist directory (exclude API routes)
-const distPath = path.join(__dirname, 'dist');
-if (fs.existsSync(distPath)) {
-  app.use((req, res, next) => {
-    // Skip static file serving for API routes
-    if (req.path.startsWith('/api/')) {
-      return next();
-    }
-    express.static(distPath)(req, res, next);
-  });
-
-  // Financial KPI Summary endpoint
-  app.get('/api/financial/kpi-summary', async (req, res) => {
-    logger.info('📊 KPI summary data requested');
-    
-    try {
-      // Attempt to get real data from external APIs
-      if (!prisma) {
-        return res.status(503).json({
-          success: false,
-          error: 'Database connection unavailable',
-          message: 'Unable to retrieve KPI data - database not connected',
-          timestamp: new Date().toISOString(),
-          userAction: 'Contact system administrator to check database configuration'
-        });
-      }
-
-      // Try to get data from database or external APIs
-      // This should integrate with real data sources like Xero, Shopify, etc.
-      return res.status(503).json({
-        success: false,
-        error: 'External API integration required',
-        message: 'KPI summary requires connection to external financial systems (Xero, Shopify, etc.)',
-        timestamp: new Date().toISOString(),
-        userAction: 'Configure external API integrations to view KPI data',
-        requiredIntegrations: ['Xero API', 'Shopify API', 'Unleashed API']
-      });
-
-    } catch (error) {
-      logger.error('Failed to fetch KPI summary:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Internal server error',
-        message: 'Unable to process KPI summary request',
-        timestamp: new Date().toISOString()
-      });
-    }
-  });
-
-  // Product Sales Performance endpoint
-  app.get('/api/sales/product-performance', async (req, res) => {
-    logger.info('📈 Product sales data requested');
-    const period = req.query.period || 'year';
-    
-    try {
-      // Attempt to get real sales data from Shopify/Amazon APIs
-      if (!prisma) {
-        return res.status(503).json({
-          success: false,
-          error: 'Database connection unavailable',
-          message: 'Unable to retrieve sales data - database not connected',
-          timestamp: new Date().toISOString(),
-          userAction: 'Contact system administrator to check database configuration'
-        });
-      }
-
-      // Check for external API integrations
-      return res.status(503).json({
-        success: false,
-        error: 'Sales data integration required',
-        message: 'Product sales performance requires connection to e-commerce platforms',
-        timestamp: new Date().toISOString(),
-        userAction: 'Configure Shopify, Amazon, or other sales platform integrations',
-        requiredIntegrations: ['Shopify API', 'Amazon SP-API'],
-        requestedPeriod: period
-      });
-
-    } catch (error) {
-      logger.error('Failed to fetch sales performance data:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Internal server error',
-        message: 'Unable to process sales performance request',
-        timestamp: new Date().toISOString()
-      });
-    }
-  });
-
-  // P&L Analysis endpoint
-  app.get('/api/financial/pl-analysis', async (req, res) => {
-    logger.info('💼 P&L analysis data requested');
-    const period = req.query.period || 'year';
-    
-    try {
-      // Attempt to get real P&L data from financial systems
-      if (!prisma) {
-        return res.status(503).json({
-          success: false,
-          error: 'Database connection unavailable',
-          message: 'Unable to retrieve P&L data - database not connected',
-          timestamp: new Date().toISOString(),
-          userAction: 'Contact system administrator to check database configuration'
-        });
-      }
-
-      // Check for financial system integrations
-      return res.status(503).json({
-        success: false,
-        error: 'Financial system integration required',
-        message: 'P&L analysis requires connection to accounting and financial systems',
-        timestamp: new Date().toISOString(),
-        userAction: 'Configure Xero, QuickBooks, or other accounting system integrations',
-        requiredIntegrations: ['Xero API', 'QuickBooks API', 'SAP integration'],
-        requestedPeriod: period
-      });
-
-    } catch (error) {
-      logger.error('Failed to fetch P&L analysis data:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Internal server error',
-        message: 'Unable to process P&L analysis request',
-        timestamp: new Date().toISOString()
-      });
-    }
-  });
-
-  // Regional Performance endpoint
-  app.get('/api/regional/performance', async (req, res) => {
-    logger.info('🌍 Regional performance data requested');
-    
-    try {
-      // Attempt to get real regional data from multiple sources
-      if (!prisma) {
-        return res.status(503).json({
-          success: false,
-          error: 'Database connection unavailable',
-          message: 'Unable to retrieve regional data - database not connected',
-          timestamp: new Date().toISOString(),
-          userAction: 'Contact system administrator to check database configuration'
-        });
-      }
-
-      // Check for regional data integrations
-      return res.status(503).json({
-        success: false,
-        error: 'Regional data integration required',
-        message: 'Regional performance requires connection to multiple regional data sources',
-        timestamp: new Date().toISOString(),
-        userAction: 'Configure regional sales, inventory, and financial system integrations',
-        requiredIntegrations: [
-          'Regional Shopify stores',
-          'Regional Amazon marketplaces', 
-          'Regional Xero entities',
-          'Regional inventory systems'
-        ]
-      });
-
-    } catch (error) {
-      logger.error('Failed to fetch regional performance data:', error);
-      return res.status(500).json({
-        success: false,
-        error: 'Internal server error',
-        message: 'Unable to process regional performance request',
-        timestamp: new Date().toISOString()
-      });
-    }
-  });
-
-  // SPA fallback - must be last AND must exclude API routes
-  app.get('*', (req, res) => {
-    // Don't serve index.html for API routes
-    if (req.path.startsWith('/api/')) {
-      res.status(404).json({
-        error: 'API endpoint not found',
-        path: req.path,
-        method: req.method,
-        timestamp: new Date().toISOString()
-      });
-    } else {
-      res.sendFile(path.join(distPath, 'index.html'));
-    }
-  });
-} else {
-  logger.warn('dist directory not found - static files will not be served');
-}
-
-// Error handling middleware
-app.use((err, req, res, next) => {
-  logger.error('Server error', err);
-  res.status(500).json({
-    error: 'Internal server error',
-    message: NODE_ENV === 'development' ? err.message : undefined
-  });
-});
-
-// Start server
-httpServer.listen(PORT, () => {
-  logger.info(`
-========================================
-SERVER STARTED SUCCESSFULLY
-========================================
-URL: http://localhost:${PORT}
-Health: http://localhost:${PORT}/health
-API Status: http://localhost:${PORT}/api/status
-Environment: ${NODE_ENV}
-Branch: ${BRANCH}
-Database: ${dbConnected ? 'Connected' : 'Not connected'}
-========================================
-  `);
-});
-
-// Graceful shutdown
-process.on('SIGTERM', async () => {
-  console.log('SIGTERM received, shutting down gracefully...');
-
-  // Close database connection
-  if (prisma) {
-    await prisma.$disconnect();
-  }
-
-
-  // Close HTTP server
-  httpServer.close(() => {
-    console.log('Server closed');
-    process.exit(0);
-  });
-});
-
-// Handle uncaught exceptions
-process.on('uncaughtException', (error) => {
-  logger.error('Uncaught exception', error);
-  process.exit(1);
-});
-
-process.on('unhandledRejection', (error) => {
-  logger.error('Unhandled rejection', error);
-  process.exit(1);
-});
-
-export default app;
